+++ menu.py
@@ -1,13 +1,10 @@
 import pygame
 
 import sys
-from game import Game
-from fighter import Fighter
 from button import Button
 from constants import MENU_ORANGE, MENU_WHITE, WHITE
-from characters import characters
-from helper import draw_bg
-
+from helper import draw_bg, get_font
+from characterselect import run_select
 
 
 pygame.init()
@@ -22,80 +19,11 @@ play_rect = pygame.image.load("assets/Menu/PlayRect.png")
 options_rect = pygame.image.load("assets/Menu/OptionsRect.png")
 quit_rect = pygame.image.load("assets/Menu/QuitRect.png")
 
-#CHARACTER SELECTION CROPS
-elvenWarrior_crop = pygame.image.load("assets/FantasyWarrior/crop.png")
-darkWizard_crop = pygame.image.load("assets/EvilWizard/crop.png")
-
-# clock = pygame.time.Clock()
-
-def select_character():
-    p1_selected = False
-
-    while True:
-        draw_bg(menu_bg, screen)
-        mouse_pos = pygame.mouse.get_pos()
-
-        SELECT_TEXT = get_font(100).render("SELECT YOUR FIGHTERS", True, MENU_ORANGE)
-        SELECT_RECT = SELECT_TEXT.get_rect(center=(SCREEN_WIDTH / 2, 100))
-
-        #FIGHTER CARD 
-        elvenWarrior_button = Button(elvenWarrior_crop, (SCREEN_WIDTH / 2 - 20, 300), "", get_font(32), WHITE, WHITE)
-        darkWizard_button = Button(darkWizard_crop, (SCREEN_WIDTH / 2 + 20, 300), "", get_font(32), WHITE, WHITE)
-
-        screen.blit(SELECT_TEXT, SELECT_RECT)
-
-        for button in [elvenWarrior_button, darkWizard_button]:
-            button.changeColor(mouse_pos)
-            button.update(screen)
-
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                pygame.quit()
-                sys.exit()
-            if event.type == pygame.MOUSEBUTTONDOWN:
-                if elvenWarrior_button.checkForInput(mouse_pos):
-                    #query database
-                    x = characters[0]
-                    ew_sheet = x["sheet"]
-                    ew_animations_steps = x["animationSteps"]
-                    ew_soundfx = x["soundfx"]
-                    ew_volume = x["volume"]
-                    ew_data = [x["size"], x["scale"], x["offset"]]
-                    if p1_selected == False:
-                        f1 = Fighter(1, 200, 310, False, ew_data, ew_sheet, ew_animations_steps, ew_soundfx, ew_volume)
-                        p1_selected = True
-                        elvenWarrior_button.text_input = "P1"
-                    else:
-                        f2 = Fighter(2, 700, 310, True, ew_data, ew_sheet, ew_animations_steps, ew_soundfx, ew_volume)     
-                        return f1,f2
-                if darkWizard_button.checkForInput(mouse_pos):
-                    x = characters[1]
-                    dw_sheet = x["sheet"]
-                    dw_animations_steps = x["animationSteps"]
-                    dw_soundfx = x["soundfx"]
-                    dw_volume = x["volume"]
-                    dw_data = [x["size"], x["scale"], x["offset"]]
-                    if p1_selected == False:
-                        f1 = Fighter(1, 200, 310, False, dw_data, dw_sheet, dw_animations_steps, dw_soundfx, dw_volume)
-                        p1_selected = True
-                        darkWizard_button.text_input = "P1"
-                    else:
-                        f2 = Fighter(2, 700, 310, True, dw_data, dw_sheet, dw_animations_steps, dw_soundfx, dw_volume)
-                        return f1,f2
-
-        pygame.display.update()	
 
 #FUNCTIONS
-def get_font(size):
-    return pygame.font.Font("assets/Fonts/8bit.ttf", size)
-
-
-
-def play():
-    fighter_1, fighter_2 = select_character()
-    pygame.display.set_caption("Fight!") 
-    game = Game(fighter_1, fighter_2, screen)
-    game.run_game()
+def select_characters():
+    run_select(screen, menu_bg)
+    
 
 def options():
     pass
@@ -128,7 +56,7 @@ def main_menu():
                 sys.exit()
             if event.type == pygame.MOUSEBUTTONDOWN:
                 if PLAY_BUTTON.checkForInput(mouse_pos):
-                    play()
+                    select_characters()
                 if OPTIONS_BUTTON.checkForInput(mouse_pos):
                     options()
                 if QUIT_BUTTON.checkForInput(mouse_pos):
